refactor: XPカウントアップ処理をフック化
